Refactor GuzzleHttpClientAspect
